Fix multiple city in event edit
I added the delete of all the cities linked to an event when i edit an event.

Before when a city existed and an event had a city; when we modify it it'll add to the previous one so the event has now 2 city.
that was what i wanted at the beginning but figured i wan't it to be only in one city as it's an event or a session.