debug filter and added range price
Debug of the filtration by direction asc and desc
it was caused by the knp paginator bundle.
Fixed it by adding a knp_paginator.yaml and disabling the sort field name options.

Added price range slide on the left side menu for filter
can't find a way to make it always shown it disappear everytime a query is made